Correct bar width for stacked bar
Since a stacked bar is displayed as a single vertical bar with segments representing the different datasets, the width of each bar corresponds to the entire bar width available and does not depend on the number of datasets. The x-coordinate of the bar is at the beginning of the section available for that bar.

Currently, with multiple datasets, the bar is not given its full width.